refactor(billing): move computeBillingCycle to repository as private method
The function is only used within gormBillingUsageRepository, so it
belongs there as an unexported function rather than in the entities
package.
